+++ src/main/java/com/inq/wishhair/wesharewishhair/review/domain/Review.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.review.domain;
 
 import com.inq.wishhair.wesharewishhair.auditing.BaseEntity;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
-import com.inq.wishhair.wesharewishhair.likereview.domain.LikeReview;
+import com.inq.wishhair.wesharewishhair.review.domain.likereview.domain.LikeReview;
 import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
 import com.inq.wishhair.wesharewishhair.review.enums.Score;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
