+++ src/main/java/com/inq/wishhair/wesharewishhair/review/service/ReviewService.java
@@ -3,8 +3,8 @@ package com.inq.wishhair.wesharewishhair.review.service;
 import com.inq.wishhair.wesharewishhair.global.consts.Condition;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyleRepository;
-import com.inq.wishhair.wesharewishhair.likereview.domain.LikeReview;
-import com.inq.wishhair.wesharewishhair.likereview.domain.LikeReviewRepository;
+import com.inq.wishhair.wesharewishhair.review.domain.likereview.domain.LikeReview;
+import com.inq.wishhair.wesharewishhair.review.domain.likereview.domain.LikeReviewRepository;
 import com.inq.wishhair.wesharewishhair.photo.PhotoStore;
 import com.inq.wishhair.wesharewishhair.photo.entity.Photo;
 import com.inq.wishhair.wesharewishhair.review.domain.Review;
